[logging] Move package into apimachinery
