+++ codegen/cuekind/parser_test.go
@@ -1,2 +1,1 @@
 package cuekind
-
