+++ codegen/templates/app/app.tmpl
@@ -4,8 +4,8 @@ import (
     "context"
     "fmt"
 
+	"github.com/grafana/grafana-app-sdk/apimachinery/logging"
 	"github.com/grafana/grafana-app-sdk/app"
-	"github.com/grafana/grafana-app-sdk/logging"
 	"github.com/grafana/grafana-app-sdk/simple"
 
 	generated "{{.Repo}}/{{.CodegenPath}}"
