+++ codegen/templates/app/watcher.tmpl
@@ -4,7 +4,7 @@ import (
     "context"
     "fmt"
 
-    "github.com/grafana/grafana-app-sdk/logging"
+    "github.com/grafana/grafana-app-sdk/apimachinery/logging"
     "github.com/grafana/grafana-app-sdk/operator"
 	"github.com/grafana/grafana-app-sdk/resource"
 	"go.opentelemetry.io/otel"
