+++ codegen/templates/plugin/plugin.tmpl
@@ -8,7 +8,7 @@ import (
 
 	"{{.Repo}}/pkg/plugin/secure"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
-	"github.com/grafana/grafana-app-sdk/logging"
+	"github.com/grafana/grafana-app-sdk/apimachinery/logging"
 	"github.com/grafana/grafana-app-sdk/plugin/kubeconfig"
 	"github.com/grafana/grafana-app-sdk/plugin/router"
 )
