+++ operator/reconciler.go
@@ -9,7 +9,7 @@ import (
 	"go.opentelemetry.io/otel/codes"
 	"k8s.io/utils/strings/slices"
 
-	"github.com/grafana/grafana-app-sdk/logging"
+	"github.com/grafana/grafana-app-sdk/apimachinery/logging"
 	"github.com/grafana/grafana-app-sdk/resource"
 )
